refactor (PurchaseAmount) : Amount를 PurchaseAmount로 리팩토링 한다
